// dine-in-restaurant-portal/src/routers/type.ts
export enum RouteLinks {
    LOGIN = '/login',
    DASHBOARD = '/',
    ANALYTICS = '/analytics',
    MENU_MANAGE = '/menu_manage',
    MENU_LIST = '/menu_list',
    ORDERS = '/orders',
    QRCODE = '/qrcode',
    INVOICES = '/invoices',
    USERS = '/users',
    KITCHEN = '/kitchen',
    NOTIFICATIONS = '/notifications',
    SETTINGS = '/settings',
    CASHIER = '/cashier',
    PAYMENTS = '/payment',
    TABLES = '/tables',
    UNAUTHORIZED = '/unauthorized',
    SUBSCRIPTIONS = '/subscriptions',
    WALLET = '/wallet',
    WAITER_CONFIRMATION = '/waiter_confirmation',
    VIEW_TABLE_ONGOING_ORDERS = '/view_table_ongoing_orders'
}